update atsu to multi atsu

+++ task1/processinglib/src/histogram.cpp
@@ -57,6 +57,6 @@ void equalized_image(Image &inImage, int histogram_equalization_formula[]) {
 
 void getNormalizedHist(int *hist, double *normalizedHist, int histSize, int pixelsNo) {
     for (int i = 0; i < histSize; i++) {
-        normalizedHist[i] = (double) hist[i] / (double) pixelsNo;
+        normalizedHist[i] = hist[i] / (double) pixelsNo;
     }
 }
